mcp_servers/requirements.txt
# Core MCP Dependencies
mcp>=1.0.0
anthropic>=0.34.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Data Processing
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
akshare>=1.12.0

# Database Connections
redis>=5.0.0
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0

# Vector Database
chromadb>=0.4.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Logging & Monitoring
loguru>=0.7.0
python-dotenv>=1.0.0

# Async Support
asyncio-mqtt>=0.13.0
aioredis>=2.0.0
aiofiles>=23.2.0

# JSON Schema & Validation
jsonschema>=4.19.0